(PUP-3801) Fixed issue with HP-UX Trusted Computing where password expirations were not being updated.
When I modified the HP-UX Password Provider in 2013, the issue was not
found until months later when the accounts used for testing exceeded the
system defined 90 day password expiration. The password hash for the new
password was getting pushed. When logging in with the new password,
HP-UX Trusted Computing was showing a successful password change, but it
was not resetting the password age back to 0 days. 

This fix corrects that behavior.