(PUP-3912) Remove unused respond_to? method
Under rspec 3, with this respond_to? in place, specs using:
    expect(stuff).to_not have_issue(issue)
were failing like so:

  1) validating 4x should not raise error for variable name with underscore first in first name segment
     Failure/Error: expect(validate(fqn('_aa').var())).to_not have_issue(Puppet::Pops::Issues::ILLEGAL_VAR_NAME)
     ArgumentError:
       wrong number of arguments (2 for 1)
     # ./spec/unit/pops/factory_rspec_helper.rb:48:in `respond_to?'
     # ./spec/unit/pops/validator/validator_spec.rb:34:in `block (2 levels) in <top (required)>'

I got stymied tracking this down, but it appears that the respond_to? in the
matcher got confused with rspec internal respond_to? metaprogramming logic (I think).

It also appears that this respond_to? isn't used for anything, so I just
removed it.