(PUP-3912) Remove spec for loading features from disk
This spec test relied on lib/puppet/util/feature.rb catching
a method_missing; rspec 2 allowed this but rspec 3 intercepts
this for rspec purposes.

We could reimplement this test in a couple ways, but this
functionality (loading features from disk) is well-tested several
other ways since the core puppet code relies on this ability to
function (and this test was written very early on).

So this commit simply removes the spec as redundant and not worth
reimplementing for rspec 3.