(PUP-3806) Add missing support for interpolation in collection
The support for double quoted (interpolated) strings were
missing in the new implementation of collectors introduced in
PUP-2906.This was not detected since there is a parser optimization
that treats double quoted strings that does not require
interpolation as literal strings and there were no tests
that included double quited strings with interpolation of an expression.

This commit adds support for concatenated string (i.e. interpolation).