(PUP-3526) Define routes in one place
Previously, all of Puppet's routes were defined separately in the Webrick and
Rack REST interfaces. This was okay when all that was being registered was the
v1 and v2 routes. However, with the CA routes now separated from the master
routes, there is quite a bit more logic there, and it is unwieldy (and hard to
test) with it defined in both places.

This commit defines the CA and master routes in one place, so that only
ca_routes and master_routes have to be registered for each webserver.