(PUP-1343) Use chkconfig --add to ensure K-symlinks get installed
On RHEL/CentOS 5 systems, the use of chkconfig on to enable a service
does not create the K-symlinks in /etc/rc.d to shut down the service.
Running chkconfig --add before chkconfig on fixes this issue, and
should be harmless on the unaffected OS versions.

The rspec unit test has been updated, and a new acceptance test
included to validate the bugfix.