(PUP-3254) Removing Modulefile functionality
Before this commit, the use of Modulefile for containing the metadata
for a module was deprecated, but still usable. This commit removes that
functionality completely, removes the deprecation warnings for
Modulefile, changes/removes the spec and acceptance tests for
Modulefile, and issues a warning that Modulefile will be ignored as a
special file.