(PUP-3684) Rename Settings.set_value to Settings.patch_value
This renames the deprecated set_value method to path_value,
removes the deprecation (in favor of the Settings []= method), and
clearly markes the method as being @api private.

We would like to get rid of the method completely but this requires
redesigning the handling of the setting --dns-alt-names which also
exists as --dns_alt_names in the certificate face. This construct is
left unchanged (except using the now renamed method). This makes it
easier to later refactor this code without having to go through a
deprecation period again for patch_value. See PUP-3836.