(PUP-2531) Add support to validate a local file reference
Since validate is mostly useful when developing a module and there
is no modulepath to use for looking up a template using a
modulename/templatename reference it becomes cumbersome to test
a template in a local file. While it was possible to use an absolute
reference, this becomes replacing one difficulty with another.

This commit adds logic that first checks if the given reference can
be loaded from the module path, and if that fails, tries to load it as
a file reference. This means that this now works:

puppet epp validate foo.epp
puppet epp validate ./somehwere/templates/foo.epp