(PUP-2531) Refactor error handling of epp face
The error handling and the final exit status was not handled
consistently. The Puppet Face API requires that actions produce
the string that is to be output to stdout, or that they raise an
exception and that the exception raising causes an exit status != 0
when run as an command line application.

The implementation called exit directly for some actions but not all.

This commit changes the behavior to print to a String buffer and then
returning that or raising an exception of the action should exit.

This has a small negative side effect that partial results can not be
viewed when processing multiple files; output is either produced or the
action fails. This is how the face api is designed; producing partial
result and also exiting is not possible without violating the API.