(PUP-3665) Make tagging arguments case insensitive
Tags are stored internally as downcased strings. All objects passed
as tags to the tagged? method where converted to strings prior to
comparison but they were not downcased. As a result, a query for
tags containing uppercase letters would always fail.

This commit adds the needed downcasing of the tags parameter. It
also introduces a new raw_tagged? method that assumes that whatever
tags that are passed to it already have been converted to strings
and downcased. This to allow the CollectorTransformer to do this
conversion once and then ask multiple resources if they are tagged.