(PUP-3755) Add spec test asserting that to_resource is in same env
This adds a test that asserts that the call to `to_resource` on
the resulting catalog of a call to Compiler.compile(a_node) is performed
in the environment configured to be used for the node.

This is very hard to test given scenarios that cause a problem if
the environment is not the same; such problems may be fixed by avoiding
logic that requires content in the environment.

Thus, this test is implementation level; checking that the call takes
place when the Puppet context has the expected state
(current_environment is 'production' rather than the initial '*root*'
environment).