(PUP-3844) Rename the MSI
This change is necessary because we are renaming the puppet installer
on windows. Since we want to maintain the ability to build older
versions of the puppet installer with the old name, this new var will
allow us to know when we expect the new name, and when we expect the old
name.