(PUP-3850) Only execute inner group on windows AND with elevated privileges
Previously, one of the inner example groups executed on non-windows
platforms when running as root. This was because the outer rspec filter
restricted the test to windows, but the inner filter restricted it to
running as root, and nested rspec filters are OR'ed, not AND'ed!

This commit ensures the inner example group is only executed when run on
windows and with elevated privileges as was originally intended.