(PUP-3850) Expect -M when executing useradd on RedHat
Previously, the tests for the useradd provider, and the openbsd
provider that extended it, failed on osfamily RedHat, because the
providers added -M when the homedir was not managed.

This commit changes the useradd_spec to test both cases, when we're on
RedHat and not. It also stubs the osfamily fact in openbsd_spec, so
that we actually test the useradd behavior on openbsd.