(PUP-3850) Set the defaultprovider for the init service test
Previously, the init_spec test failed on Fedora, because puppet loaded
each service provider, checking to see which was the default. Due to
how providers are confined, it caused Facter to load various facts,
interfering with the stubbed `File.directory?` method.

This commit specifies the defaultprovider to short circuit that process.
It also clears the defaultprovider, similar to what was done in
commit 6d2eeaaee.