(PUP-3801) HP-UX Trusted Computing Password Age Update
HP-UX Trusted Computing does not change the password age when doing an
update to the password hash using /usr/sam/lbin/usermod.sam. This change
corrects that behavior and sets the password age back to 0 days old
after it has been modified. If the HP-UX box is not running Trusted
Computing it only change the password.