(PUP-3002) Improve error message for invalid parameter name
The error message emitted when an attribute operation used
an attribute name that was invalid did not read well and caused
confusion since it did not clearly indicate that part of the message was
the name of the parameter - e.g. "invalid parameter hash" which can be
read as an invalid "parameter hash" (confusing when actually setting all
parameters from a hash and having a parameter named 'hash' in the hash).

The message is now changed to say "Invalid parameter: '<name>'". As a
consequence spec tests are updated to match this message.