(PUP-3548) Allow 4x functions to call 3x functions
The call_function() method of the Puppet::Pops::Functions::Function
class did not make any attempt to find functions that had been
declared using the 3x function API. This commit adds the following:

Proper load and execute:
The call_function() method will now consult the 3x Functions class
to find (and optionally load) the function from there when it fails
to find it using the 4x loader.

Other improvements:
The initialize method in Puppet::Pops::Evaluator::Runtime3Support
was removed in favor of accessors that creates the needed visitors.
This was done to get rid of code in the EvaluatorImpl initializer
that otherwize would need to be copied into the ArgsConverter class
as well.