(PUP-3890) Fix mtime/ctime checksum type
Previously file resources compared the ctime/mtime of targets
against the source by comparing the checksum strings to see if they're
equal. That would always update the target file, because it would be
given a new modification time that doesn't match the source.

This change makes file contents that use mtime or ctime for
checksum_type compare to ensure the target mtime/ctime is newer than
the source. The target will only be updated if it's older than the
source. This change doesn't affect writing from contents because it's
an error to use ctime/mtime with a content attribute.

Removes the checksum_type method from contents, because it's now just an
alias back to the resource. Better to query the resource directly.