(PUP-3855) Restore ability for auth.conf to restrict on environment
The PR for PUP-3644 moved the `check_authorization` call for
indirected routes to a spot before the environment for a request is
translated into a `Puppet::Node::Environment` instance in the request
params.  This caused any restrictions on environment made in an
auth.conf file to inappropriately reject all requests matching the
rule path, even ones that should have been allowed.

This PR moves the `check_authorization` call back to a spot in the
code after the translation to an `Puppet::Node:::Environment` instance
occurs.  This change restores the ability for an environment value in
the auth.conf to appropriately allow/deny access for requests.