(maint) Update CI Setup for AIO Testing
This implements a new pre-suite that installs the puppet-agent (all in
one) package and add a new Rake task "ci:test:aio" to utilize the new
pre-suite.

note: there are currently failures in the pre-suite for AIO because of
issues between the AIO package and the puppet server. These are expected
to be ironed out with the dev release of puppet-server 2.0