(PUP-3897) Add acceptance: ensure new environment picked up by agent
in PUP-3755 there are two scenarios that lead to a newly classified
agent not picking up its new environment.  This test should catch both
of the cases that aren't tested by spec.
This test does not use the environment helpers as they circumvent most
of the functionality in this test such as hard wiring to test
environment and specifying agent environment on commandline.  We need to
classify a node's environment from master.