(maint) Improve the usefulness of the BindingsModelDumper
The BindingsModelDumper is a debugging aid for the Puppet Binder. The
format was both hard to read, and could not be directly used on the top
level object of type Injector (which is really the object that is
visible in logic using the binder).

This commit adds the ability to simply dump the Injector to get output
of the effective entries (the resulting composition of bindings) in a
less daunting format.

Example usage in debugger:
```
> puts Puppet::Pops::Binder::BindingsModelDumper.new.dump(injector)
```
Using `puts` outputs the resulting string with new lines.

The dumper is by far complete, and it is expected that future debugging
will require it to be modifid to make it even more useful.