(PUP-3900) Make loading of bindings resolve classes in current env
Before this commit, the resolution of a class_name to instantiated
Class used an autoloader that ultimately ended up using the default
environment's location and module path.

As a consequence of fixing this, the bindings_composer_spec failed
due to current_environment not being set.

As a consequence of errors in class name to path resolution, naming
errors in fixtures went unnoticed. This commit changes the use of
'Puppetx' to the expected 'PuppetX', and changes path from 'puppetx', to
'puppet_x'.