(PUP-3900) Add spec for data providers from a module
This adds a spec test that adds both an env and a module data provider
to the set of providers. It has an environment called 'sample' that
uses the env data provider with the same name, and a module data
provider also with the same name.

The fixture used also serves as an example configuration for someone
that wants to write providers.