(PUP-3924) Remove the hiera "puppet_backend" from puppet
This odd backend allowed hiera lookups to take place from a puppet
catalog. A very strange feature.

This removes the "puppet_backend" from puppet's code base. Not only was
it odd, it also creates a circular dependency between the hiera and
puppet respositories since puppet depends on a particular version of
hiera source which blocks progress. (Hiera has to be released before
puppet can be updated in order to run tests, yet when running them from
source to test the combination, puppet needs to be updated).

With the backend gone, there is a normal dependency on hiera's API only.