(PUP-5094) Have the compiler evaluate applications
We defer evaluation of applications until after we are done evaluating the
usual stuff applying to a node. The evaluate_applications method then makes
a pass over all application instances we found during 'normal' evaluation.

For each application instance, determine which components apply to the current
node and evaluate those components. This is done by evaluating the application
instance as a whole, but skipping further evaluation (in add_resource) of any
component that isn't mapped to the current node.

This will probably fail if an application definition contains components nested
at a lower level.  IOW, this code assumes that all components are at the top
level within an application definition and not nested within other
defined types.