(PUP-2916) If reposdir is set, replace the default directories instead of appending
When a custom reposdir is present in the yum.conf file, yum will only
read files from that directory/directories (plus yum.conf).

Previously, the yumrepo provider was reading all files from all
directories in the reposdir array including the default directory, even
though they would have been ignored by yum. This was an issue if the
same repo existed in both the default directory and the custom
directory. This change just replaces the default directories instead of
appending to them if reposdir is set.