(PUP-3678) Correct and unify module name validation
Before this commit, we were checking the validity of a module name in
multiple places. We were also only excluding . and .. from the list
of modules to create. This permitted directories in the modulepath like
.git, .svn, and .DS_store to cause errors and warnings. This commit
consolidates the module name checking and excludes those types of
directories. It also enforces a name validity check more in line with
the Forge validator. Specifically, it also enforces no namespaced
module directory names (ex. author-modulename) in the modulepaths.