(PUP-3863) Move 3x hiera functions to 4x API
This commit moves all four hiera functions from 3x to 4x. It also moves the
parameter validation from the HieraPuppet class into the 4x function dispatch.

Tests are added to verify the validity of the new functions (and invalidity of
the old).

Since three functions are identical in all aspects except the merge type
argument sent to hiera, and since the merge type can be derived from the name
of the function, an approach was taken where these three functions share the
same class definition block. The hiera_include is separate though, since its
logic differs.