(PUP-3934) Propagate environment to Resource initializer
Prior to this commit, if a Resource was created as 'Bar[instance]',
it would fail an attemt to fetch the resource type from its
environment. The environment would not be set and thus fallback
to whatever the :current_environment referred to. This made
the test spec/integration/parser/ruby_manifest_spec.rb fail.

Here, the environment is passed from the catalog to the Resource
initialize methdo in the attributes hash and properly assigned
before used.