(maint) Create master module directories during git setup
Commit dc86396c modified `acceptance/tests/modules/build/build_basic.rb`
so that it no longer created the dist and site module directories on the
master. It turns out that later tests relied on that behavior and
fail if that directory does not exist, e.g.
`tests/modules/already_installed.rb`.

The commit did not affect `package` type acceptance, because the puppet
server package creates those directories. However, it did cause `git`
type acceptance to fail, such as is used on Windows. It would have also
caused Solaris to fail, except that we don't run PMT acceptance on
Solaris currently.

This commit updates the setup step used to configure the puppet master,
but only for `git`. The other types are unaffected.

Note several other module acceptance tests still attempt to `mkdir -p`
the site and dist module directories, and those calls should be removed
at some point.