(PUP-3960) Verify rpms installed for rpm_util helper
This commit adds a 'setup' method to the Puppet::Acceptance::RpmUtils module.
This method installs the rpms required by the other methods in the module.
Currently, these rpms are createrepo and rpm-build.

The acceptable_exit_codes have been expanded in clean_rpm to allow this method
to gracefully continue if the given rpms are not on the system when the
clean_rpm method is called.

The new 'setup' method is also added to the beginning of the setup_rpm and
send_rpm methods to ensure that their requirements are satisfied.