(PUP-3969) Use Beaker puppet method for host 'user' and 'group'
This commit updates puppet acceptance tests and their helper methods to use
the Beaker `host.puppet['user']` and `host.puppet['group']` method in order to
introspect the proper puppet user and group.

Previously, these tests and helpers were relying on the `host` hash as defined
by Beaker to determine these values. This method is no longer proper procedure
for determining these values and was resulting in test failures in the
acceptance testing of puppet against the All In One agent as well as
failures in the acceptance testing of Beaker.