(PUP-3968) Cache the `zypper list-updates` output
The zypper provider calls `zypper list-updates` in order to get a list of the
packages that have newer versions. Then it checks if the resource name matches
any of the items in the list, and then applies the newer version. The problem
with the above though is that it runs list-updates for every single package
instance, while it should be run only once, as it provides all the packages
that have newer versions available.

This commit caches the list-updates output and puts its contents in a hash, so
that each package resource will query this instead of zypper directly

Signed-off-by: Adrien Thebo <adrienthebo@somethigsinistral.net>