(PUP-3948) Add helper class for type assertion
The Puppet::Pops::Types::TypeAsserter class added by this commit
makes asserting that an instance is of a given type, or that a type
is assignable to another type, and raising an exceptions for the
failing cases, a one-liner.

Types are frequently checked in various places of the code and
exceptions are raised when the such check fails. A small helper
class that can perform this simple task rather than raising different
types of exceptions with varying messages is needed to achieve
consistency as the number of such checks grow.